# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/students": {
      "get": {
        "operationId": "StudentsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Serbest metin arama",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "institutionId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scholarshipProgramId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "required": false,
            "in": "query",
            "description": "Grupta su an aktif olanlar",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gender",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["FEMALE", "MALE"],
              "type": "string"
            }
          },
          {
            "name": "universityYear",
            "required": false,
            "in": "query",
            "description": "0 = hazirlik",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "default": "active",
              "enum": ["active", "withdrawn", "all"],
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "name",
              "enum": ["name", "studentNumber", "enrollDate", "createdAt"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["students"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "StudentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["students"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/students/export": {
      "get": {
        "operationId": "StudentsController_export",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Serbest metin arama",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "institutionId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scholarshipProgramId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "required": false,
            "in": "query",
            "description": "Grupta su an aktif olanlar",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gender",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["FEMALE", "MALE"],
              "type": "string"
            }
          },
          {
            "name": "universityYear",
            "required": false,
            "in": "query",
            "description": "0 = hazirlik",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "default": "active",
              "enum": ["active", "withdrawn", "all"],
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "name",
              "enum": ["name", "studentNumber", "enrollDate", "createdAt"],
              "type": "string"
            }
          },
          {
            "name": "format",
            "required": true,
            "in": "query",
            "schema": {
              "enum": ["csv", "excel", "pdf"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["students"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/students/import-template": {
      "get": {
        "operationId": "StudentsController_importTemplate",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["students"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/students/import": {
      "post": {
        "operationId": "StudentsController_import",
        "parameters": [
          {
            "name": "institutionId",
            "required": false,
            "in": "query",
            "description": "Sadece sistem yoneticisi icin zorunlu",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["students"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/students/{id}": {
      "get": {
        "operationId": "StudentsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["students"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "StudentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["students"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "StudentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["students"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/students/{id}/withdraw": {
      "post": {
        "operationId": "StudentsController_withdraw",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["students"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/students/{id}/reinstate": {
      "post": {
        "operationId": "StudentsController_reinstate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["students"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/students/{id}/groups": {
      "get": {
        "operationId": "StudentsController_groups",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["students"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/students/{id}/group-transfers": {
      "post": {
        "operationId": "StudentsController_transfer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupTransferDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["students"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/students/{id}/attendance": {
      "get": {
        "operationId": "StudentsController_attendance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-10-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2027-06-30",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["students"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/students/{id}/qr-token": {
      "get": {
        "operationId": "StudentsController_qrToken",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["students"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Öğrenci Takip Sistemi API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateStudentDto": {
        "type": "object",
        "properties": {
          "institutionId": {
            "type": "string",
            "description": "Sadece sistem yoneticisi icin zorunlu"
          },
          "studentNumber": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": ["FEMALE", "MALE"],
            "description": "Cinsiyeti tanimli yurtta zorunlu"
          },
          "phone": {
            "type": "string"
          },
          "university": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "universityYear": {
            "type": "number",
            "description": "0 = hazirlik"
          },
          "scholarshipProgramId": {
            "type": "string"
          },
          "guardianName": {
            "type": "string"
          },
          "guardianPhone": {
            "type": "string"
          },
          "guardianEmail": {
            "type": "string"
          },
          "enrollDate": {
            "type": "string",
            "example": "2026-10-01"
          }
        },
        "required": ["studentNumber", "firstName", "lastName", "enrollDate"]
      },
      "UpdateStudentDto": {
        "type": "object",
        "properties": {
          "studentNumber": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": ["FEMALE", "MALE"],
            "description": "Cinsiyeti tanimli yurtta zorunlu"
          },
          "phone": {
            "type": "string"
          },
          "university": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "universityYear": {
            "type": "number",
            "description": "0 = hazirlik"
          },
          "scholarshipProgramId": {
            "type": "string"
          },
          "guardianName": {
            "type": "string"
          },
          "guardianPhone": {
            "type": "string"
          },
          "guardianEmail": {
            "type": "string"
          },
          "enrollDate": {
            "type": "string",
            "example": "2026-10-01"
          }
        }
      },
      "WithdrawStudentDto": {
        "type": "object",
        "properties": {
          "withdrawDate": {
            "type": "string",
            "description": "Bos ise bugun"
          }
        }
      },
      "GroupTransferDto": {
        "type": "object",
        "properties": {
          "fromGroupId": {
            "type": "string"
          },
          "toGroupId": {
            "type": "string"
          },
          "effectiveDate": {
            "type": "string",
            "example": "2026-11-01",
            "description": "Yeni grupta ilk gun"
          }
        },
        "required": ["fromGroupId", "toGroupId", "effectiveDate"]
      },
...
    }
  }
}